Merging a lot from dev to main

+++ tools/create-clients.cmd
@@ -1,4 +1,5 @@
 @echo off
+CLS
 
 SET cdir=%~dp0
 SET apiDll="bin\Debug\net6.0\TeamCloud.API.dll"
@@ -29,7 +30,15 @@ popd
 pushd %cdir%..\web
 
     echo Uninstalling teamcloud from web
-    call npm uninstall teamcloud
+    call npm uninstall teamcloud --legacy-peer-deps
+    echo.
+
+    echo Deleting web node_modules
+    IF EXIST .\node_modules\ rd /q /s .\node_modules\
+    echo.
+
+    echo Deleting package lock
+    IF EXIST .\package-lock.json del /f /q .\package-lock.json
     echo.
 
 popd
@@ -37,11 +46,11 @@ popd
 pushd %cdir%..\web\teamcloud
 
     echo [TypeScript] Deleteing old node_modules
-    rd /s /q .\node_modules
+    IF EXIST .\node_modules\ rd /q /s .\node_modules\
     echo.
 
     echo [TypeScript] Deleteing package.lock
-    del /f /q .\package-lock.json
+    IF EXIST .\package-lock.json del /f /q .\package-lock.json
     echo.
 
 popd
@@ -52,9 +61,9 @@ pushd %cdir%..\openapi
     call autorest --reset
     echo.
 
-    # echo "Generating python client"
-    # autorest --v3 python.md
-    # echo ""
+    echo "Generating python client"
+    call autorest --v3 python.md
+    echo ""
 
     echo [TypeScript] Generating client
     call autorest --v3 typescript.md
@@ -68,8 +77,16 @@ pushd %cdir%..\web\teamcloud
     call npm install
     echo.
 
+    echo [TypeScript] adding rimraf to dev dependencies
+    call npm install rimraf@^3.0.0 -D
+    echo.
+
     echo [TypeScript] Building client
-    call npm run-script build
+    call npm run build
+    echo.
+
+    echo [TypeScript] Deleteing README.md
+    IF EXIST ./README.md del /f /q ./README.md
     echo.
 
 popd
@@ -77,7 +94,7 @@ popd
 pushd %cdir%..\web
 
     echo [TypeScript] Installing temacloud to web
-    call npm install .\teamcloud
+    call npm install .\teamcloud --legacy-peer-deps
     echo.
 
 popd
